Run 'hash -r' after adding symlinks
We already had this in the activate script, but this isn't sufficient.
It's only a problem right after a symlink is added and the user wants to
use it within the same session.

In order to fix this, we need to use a shell function to intercept
invocations of cachenv because 'hash -r' needs to be run from the user's
shell.